==> notebook/00_EXTRACT_DATA.ipynb <==
## Load libraries

from bs4 import BeautifulSoup
import pandas as pd
import numpy as np
from IPython.core.display import clear_output
from warnings import warn
import os
#from tqdm import tqdm
from tqdm.notebook import trange, tqdm


import requests
import time
from requests.adapters import HTTPAdapter
from requests.packages.urllib3.util.retry import Retry

## Create functions

def requests_retry_session(
    retries=6,
    backoff_factor=0.3,
    status_forcelist=(500, 502, 504),
    session=None,
):
    session = session or requests.Session()
    retry = Retry(
        total=retries,
        read=retries,
        connect=retries,
        backoff_factor=backoff_factor,
        status_forcelist=status_forcelist,
    )
    adapter = HTTPAdapter(max_retries=retry)
    session.mount('http://', adapter)
    session.mount('https://', adapter)
    return session

def get_response(url):
    t0 = time.time()
    try:
        response = requests_retry_session().get(
            url,
            timeout=10
        )
        return response
    except Exception as x:
        print('It failed :(', x.__class__.__name__)
    else:
        print('It eventually worked', response.status_code)
    finally:
        t1 = time.time()
        print('Took', t1 - t0, 'seconds')

def bs_parser_post(response):
    '''extract title, link, tags, & summaryfrom request data'''
    titles = []
    summaries = []
    links = []
    tags = []
    for p in tqdm(range(len(response))):
        page_html = BeautifulSoup(response[p].text, 'html.parser')
        containers = page_html.find_all('div', class_ = 'fusion-post-content post-content')
        #print(len(containers))
        for i in containers:
            ti =  i.h2.text
            li = i.a['href']
            ta = i.select('span')[6].get_text(strip=True)[5:]
            su = i.find('div', class_ = 'fusion-post-content-container').p.text
            titles.append(ti)
            links.append(li)
            tags.append(ta)
            summaries.append(su)
    return titles,summaries,tags,links

## Create list of responses

rlist = []
# loop over 32 times
for i in tqdm(range(1,33)):
    url = 'https://scrapsfromtheloft.com/tag/stand-up-transcripts/page/{}'.format(i)
    r = get_response(url)
    rlist.append(r)

len(rlist)

## Create Dataframe with data

titles,summaries,tags,links = bs_parser_post(rlist)

df = pd.DataFrame({
                   'raw_title': titles,
                   'summary': summaries,
                   'tags':tags,
                   'link': links})
df.info()

df.head()

## Save dataframe

data_directory_saves = os.path.join( '..','data','raw_data/')

df.to_pickle(data_directory_saves+'raw_comedy.pkl')
