==> notebook/01_EXTRACT_DATA.ipynb <==
## Imports

import pandas as pd
import numpy as np
import pickle
from warnings import warn
import os

## Load Data

file = 'raw_comedy.pkl'
data_directory = os.path.join('..','data','raw_data/{}'.format(file))
data_directory_saves = os.path.join( '..','data','raw_data/')

data_directory

df = pd.read_pickle(data_directory)
df.info()

df.head()

## Tags

df.tags.value_counts(dropna=False)

tags = df.tags.str.split(',',expand=True)
tags.head()

tags[0].value_counts()[:20]

tags[1].value_counts()

tags[2].value_counts()

tags[2].value_counts()

len(df.tags.value_counts(dropna=False))

# Remove unwanted tags

df.tags = df.tags.str.replace('Stand-up transcripts','')
df.tags = df.tags.str.replace('Gun control','')
df.tags = df.tags.str.replace('SATURDAY NIGHT LIVE','')
df.tags = df.tags.str.replace('Abortion','')
df.tags = df.tags.str.replace('Religion','')
df.tags = df.tags.str.replace(',','')

# Firstname Lastname
df.tags = df.tags.str.title()

df.tags[0]

df.tags.value_counts(dropna=False)

len(df.tags.value_counts(dropna=False))

## Rename Column

# rename column
df.rename(columns={"tags": "name"},inplace=True)

#order columns
df = df[['name','raw_title','summary','link']]

### 

# Save

df.head()

data_directory_saves = os.path.join( '..','data','raw_data/')
df.to_pickle(data_directory_saves+'raw_comedy_02.pkl')

df.raw_title.str.split(':',expand=True)

df['full_name'] = df.raw_title.str.split(':').str[0].str.title()

df.head()

df.full_name.value_counts()

df.full_name.unique().tolist()

df.name.unique().tolist()

df.head()

# Get Transcripts
