==> notebook/04_clean_data.ipynb <==
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
import os
import re

# Load data


# comedy link list
file = 'raw_comedy_03.pkl'
data_directory = os.path.join('..','data','raw_data/{}'.format(file))
df = pd.read_pickle(data_directory)
df.head()

df.head()

# Clean
## YEAR
Get eyar from `raw_title`

def get_year(data):
    ''' return year'''
    patt = r'(\d{4})' # ####
    text = data.raw_title 
    pattern = re.compile(patt) 
    matches = pattern.findall(text) 
    if len(matches) == 0: #matches = 0 
        return 0 
    else: 
        value = matches[0] 
        return value

%%time
df['year'] = df.apply(get_year,axis=1)

df.year.value_counts(dropna=False)[:10]

df.head()

## Remove from `raw_title`
`raw_title` contains name of the comedian, since we have a column that already does that, lets remove the comedian name from column

- crate a list of unique names from column `name
- fix casing on raw_title to match our list of names
- remove name from raw_title if its in our name list

# create list of names
names = list(df.name.unique())
len(names)

# fix casing 
df.raw_title = df.raw_title.str.title()

names[:10]

df.raw_title[:10]

df['title_clean'] = df['raw_title'].str.replace('|'.join(names), '',case=False)

df.head()

## remove  "full transcript"
- `title_clean` contains the several versions of the string `full transcipt`
- remove it by replacing it with an empty string

df.title_clean.value_counts()

df['title_clean'] = df.title_clean.str.replace('– Full Transcript|Full Transcript|Transcript| Stand-Up','',case=False)

df.title_clean.value_counts()

## Remove year

- since we extracted the year already,we can remove it
- regex extract 4 digits incase in parenthesis
- (####)

#yr_patt = r'\((.*?)\)'
regex_pat = re.compile(r'\((.*?)\)', flags=re.IGNORECASE)
df['title_clean'] = df.title_clean.str.replace(regex_pat, '')

df.title_clean.value_counts()

## Remove random chars
- `:`
- `-`

df['title_clean'] = df.title_clean.str.replace(':|–','')

df.title_clean.value_counts()

## Remove brackets
- `[####]`

#yr_patt = r'\((.*?)\)'
regex_pat = re.compile(r'\[(.*?)\]', flags=re.IGNORECASE)
df['title_clean'] = df.title_clean.str.replace(regex_pat, '')

# remove any whitespace from title
df['title_clean'] = df['title_clean'].str.strip()

df.title_clean.value_counts()

df.title_clean.unique().tolist()

""
# rename col
df.rename(columns={"title_clean": "title"},inplace=True)

# Save

df.head()

data_directory_saves = os.path.join( '..','data','raw_data/')
df.to_pickle(data_directory_saves+'raw_comedy_04.pkl')
