==> notebook/06_ENG_CLEAN.ipynb <==
import pandas as pd
import string
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
import re
import os
from IPython.core.display import clear_output
plt.style.use('ggplot')
sns.set(rc={'figure.figsize':(10.7,4.27)})

## Load data

# comedy link list
file = 'raw_comedy_05.pkl'
data_directory = os.path.join('..','data','raw_data/{}'.format(file))
df = pd.read_pickle(data_directory)
df.head()

data = df[['name','dob','sex','title','year','transcript']]

data.info()

data.head()

## English only

data.title.value_counts()[:10]

filter_data = data.title.str.contains('Italiano')
data[filter_data]

data = data[~filter_data]

data[filter_data]

data.info()

## Clean Text

def clean_text(data):
    #print(type(data))
    text = data.lower()
    text = re.sub('\[.*?\]', '', text)  # remove brackets & text inside
    text = re.sub('[%s]' % re.escape(string.punctuation), '', text)  # remove punctuations
    text = re.sub('\w*\d\w*', '', text)  # remove digits surounded by words
    text = re.sub('[‘’“”…]', '', text)  # remove quotes & elipses
    text = re.sub('\n', ' ', text)  # remove new line
    text = re.sub('\t', ' ', text)  # remove new line
    #text = re.sub(string.whitespace, '', text)  # remove \n\t..
    return text

%%time
data['clean_text'] = data.transcript.apply(clean_text)

data.head()

# drop transcript
data.drop(['transcript'],axis=1,inplace=True)

data.head()

## Save

data_directory_saves = os.path.join( '..','data','raw_data/')
data.to_pickle(data_directory_saves+'raw_comedy_06.pkl')
